# packages/shared/openapi/openapi-generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kyro API",
    "version": "1.0.0",
    "description": "Language-agnostic API platform for file storage and management.",
    "contact": {
      "name": "Kyro Support",
      "url": "https://kyro.io"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development"
    },
    {
      "url": "https://api.kyro.io/api/v1",
      "description": "Production"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              },
              "code": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
...
      "Organisation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "plan": {
            "type": "string",
            "enum": [
              "free",
              "pro",
              "enterprise"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Member": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "admin",
              "member"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request body",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/org": {
      "get": {
        "tags": [
          "Organisation"
        ],
        "summary": "Get current organisation",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-scope": null,
        "x-body-description": null,
        "x-response-example": "{\"id\":\"550e8400-e29b-41d4-a716-446655440000\",\"name\":\"Acme Inc\",\"slug\":\"acme-inc\",\"plan\":\"free\",\"createdAt\":\"2026-04-10T12:00:00Z\"}",
        "responses": {
          "200": {
            "description": "Organisation details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organisation"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/org/members": {
      "get": {
        "tags": [
          "Organisation"
        ],
        "summary": "List organisation members",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-scope": null,
        "x-body-description": null,
        "x-response-example": "[{\"id\":\"550e8400-e29b-41d4-a716-446655440000\",\"email\":\"admin@acme.com\",\"role\":\"owner\",\"createdAt\":\"2026-04-10T12:00:00Z\"}]",
        "responses": {
          "200": {
            "description": "List of members",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Member"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Organisation"
        ],
        "summary": "Invite new member",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "owner",
                      "admin",
                      "member"
                    ],
                    "default": "member"
                  }
                }
              }
            }
          }
        },
        "x-scope": null,
        "x-body-description": "{ \"email\": \"user@acme.com\", \"password\": \"secure123\", \"role\": \"member\" }",
        "x-response-example": "{\"id\":\"550e8400-e29b-41d4-a716-446655440000\",\"email\":\"user@acme.com\",\"role\":\"member\"}",
        "responses": {
          "201": {
            "description": "Member invited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Member"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/org/members/{id}": {
      "delete": {
        "tags": [
          "Organisation"
        ],
        "summary": "Remove member",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "x-scope": null,
        "x-body-description": null,
        "x-response-example": "204 No Content",
        "responses": {
          "204": {
            "description": "Member removed"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  }
}